requirements.txt
feedparser>=6.0.11
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
APScheduler>=3.10.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
flask>=3.0.0
